[IMP] phone_validation, crm: move onchange on phone fields
PURPOSE

Allow auto install of all SMS-based feature. Move partner onchange formatting
directly inside phone validation.

SPECIFICATIONS

Move the onchange on the phone field of contact model directly inside phone
validation to have it available in all SMS based applications.

LINKS

Task 2061765
